frontend_requirements.txt
numpy
streamlit
pillow
python-multipart
requests